+++ blocklets/ai-runtime/src/locales/zh.ts
@@ -1,0 +1,5 @@
+import flat from 'flat';
+
+export default flat({
+  openTheBot: '打开',
+});