+++ blocklets/ai-studio/src/components/file-editor/output/OutputNameCell.tsx
@@ -1,22 +1,57 @@
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
-import { OutputVariableYjs } from '@blocklet/ai-runtime/types';
+import { AssistantYjs, OutputVariableYjs } from '@blocklet/ai-runtime/types';
 import { Icon } from '@iconify-icon/react';
 import VariableIcon from '@iconify-icons/tabler/variable';
 import { Stack, TextField, TextFieldProps, Typography } from '@mui/material';
 
+import useCallAgentOutput from '../use-call-agent-output';
 import { getRuntimeOutputVariable } from './type';
 
 export default function OutputNameCell({
   depth = 0,
   output,
   TextFieldProps,
+  assistant,
+  projectId,
+  gitRef,
 }: {
   depth?: number;
   output: OutputVariableYjs;
   TextFieldProps?: TextFieldProps;
+  assistant: AssistantYjs;
+  projectId: string;
+  gitRef: string;
 }) {
   const { t } = useLocaleContext();
   const runtimeVariable = getRuntimeOutputVariable(output);
+  const { getRefOutputData, getOutputI18nName } = useCallAgentOutput({
+    projectId,
+    gitRef,
+    assistant,
+  });
+
+  const render = () => {
+    if (output.from?.type === 'output') {
+      return <Typography>{getOutputI18nName(getRefOutputData(output?.from?.id || '')?.name || '')}</Typography>;
+    }
+
+    if (runtimeVariable) {
+      return <Typography>{t(runtimeVariable.i18nKey)}</Typography>;
+    }
+
+    return (
+      <TextField
+        variant="standard"
+        fullWidth
+        hiddenLabel
+        placeholder={t('outputVariableName')}
+        {...TextFieldProps}
+        value={output.name || ''}
+        onChange={(e) => (output.name = e.target.value)}
+        onClick={(e) => e.stopPropagation()}
+      />
+    );
+  };
 
   return (
     <Stack
@@ -30,20 +65,7 @@ export default function OutputNameCell({
       }}>
       {runtimeVariable?.icon || (depth === 0 ? <Icon icon={VariableIcon} /> : undefined)}
 
-      {runtimeVariable ? (
-        <Typography>{t(runtimeVariable.i18nKey)}</Typography>
-      ) : (
-        <TextField
-          variant="standard"
-          fullWidth
-          hiddenLabel
-          placeholder={t('outputVariableName')}
-          {...TextFieldProps}
-          value={output.name || ''}
-          onChange={(e) => (output.name = e.target.value)}
-          onClick={(e) => e.stopPropagation()}
-        />
-      )}
+      {render()}
     </Stack>
   );
 }