+++ blocklets/ai-studio/src/locales/en.ts
@@ -697,6 +697,7 @@ export default flat({
   promptRequired: 'Prompt is required',
   twitter: 'Twitter',
   community: 'Community',
+  link: 'Link',
   saveAs: 'Save as',
   openingQuestions: 'Opening Questions',
   openingMessage: 'Opening Message',
@@ -743,4 +744,9 @@ export default flat({
   hideOutputSetting: 'Hide current output setting',
   activeOutputTip: 'Active output',
   hideOutputTip: 'Hide output',
+  callAgent: 'Call Agent',
+  referenceOutput: 'From {agent} Output',
+  ref: 'From',
+  refOutput: 'Call agent does not currently support custom output',
+  addParamter: 'Add Parameter',
 });