+++ packages/ai-runtime/src/types/assistant/utils.ts
@@ -7,6 +7,7 @@ import type {
   AgentYjs,
   ApiAssistantYjs,
   AssistantYjs,
+  CallAssistantYjs,
   ExecuteBlockYjs,
   FileTypeYjs,
   FunctionAssistantYjs,
@@ -23,6 +24,7 @@ import type {
   Agent,
   ApiAssistant,
   Assistant,
+  CallAssistant,
   ExecuteBlock,
   FileType,
   FunctionAssistant,
@@ -66,7 +68,7 @@ export function isAssistant(assistant: FileTypeYjs): assistant is AssistantYjs;
 export function isAssistant(assistant: FileType | FileTypeYjs): assistant is FileType | AssistantYjs {
   return (
     typeof (assistant as any).id === 'string' &&
-    ['agent', 'prompt', 'image', 'api', 'function', 'router'].includes((assistant as any).type)
+    ['agent', 'prompt', 'image', 'api', 'function', 'router', 'callAgent'].includes((assistant as any).type)
   );
 }
 
@@ -80,6 +82,12 @@ export function isRouterAssistant(file: FileType): file is RouterAssistant;
 export function isRouterAssistant(file: FileTypeYjs): file is RouterAssistantYjs;
 export function isRouterAssistant(file: FileType | FileTypeYjs): file is RouterAssistant | RouterAssistantYjs {
   return (file as any).type === 'router';
+}
+
+export function isCallAssistant(file: FileType): file is CallAssistant;
+export function isCallAssistant(file: FileTypeYjs): file is CallAssistantYjs;
+export function isCallAssistant(file: FileType | FileTypeYjs): file is CallAssistant | CallAssistantYjs {
+  return (file as any).type === 'callAgent';
 }
 
 export function isPromptAssistant(file: FileType): file is PromptAssistant;
@@ -336,6 +344,17 @@ export function fileToYjs(file: FileType): FileTypeYjs {
     };
   }
 
+  if (isCallAssistant(file)) {
+    return {
+      ...file,
+      parameters: file.parameters && arrayToYjs(file.parameters, parameterToYjs),
+      tests: file.tests && arrayToYjs(file.tests),
+      entries: file.entries && arrayToYjs(file.entries),
+      outputVariables: file.outputVariables && arrayToYjs(file.outputVariables.map(outputVariableToYjs)),
+      agents: file.agents && arrayToYjs(file.agents),
+    };
+  }
+
   return file;
 }
 
@@ -406,6 +425,17 @@ export function fileFromYjs(file: FileTypeYjs): FileType {
       entries: file.entries && arrayFromYjs(file.entries),
       outputVariables: file.outputVariables && arrayFromYjs(file.outputVariables).map(outputVariableFromYjs),
       routes: file.routes && arrayFromYjs(file.routes),
+    };
+  }
+
+  if (isCallAssistant(file)) {
+    return {
+      ...file,
+      parameters: file.parameters && arrayFromYjs(file.parameters, parameterFromYjs),
+      tests: file.tests && arrayFromYjs(file.tests),
+      entries: file.entries && arrayFromYjs(file.entries),
+      outputVariables: file.outputVariables && arrayFromYjs(file.outputVariables).map(outputVariableFromYjs),
+      agents: file.agents && arrayFromYjs(file.agents),
     };
   }