+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.3.24 (2024-7-9)
+
+- fix: should respond json result if `responseType` is not stream
+
+## 0.3.23 (2024-7-9)
+
+- test: add e2e tests
+
+## 0.3.22 (2024-7-9)
+
+- chore: add server side aigne apis
+
+## 0.3.21 (2024-7-6)
+
+- chore: agent call other agent
+
+## 0.3.20 (2024-7-5)
+
+- chore: add `runAgent` api in aigne-sdk
+
+## 0.3.19 (2024-7-5)
+
+- feat: support sharing message to ArcBlock community
+
 ## 0.3.18 (2024-7-4)
 
 - fix: failed to filter reference agent parameters in parameters